Only delete directory if it exists
Prevents potential DirectoryNotFoundException.

+++ GitUI/CommandsDialogs/FormCommit.cs
@@ -1959,7 +1959,7 @@ namespace GitUI.CommandsDialogs
                                 {
                                     File.Delete(path);
                                 }
-                                else
+                                else if (Directory.Exists(path))
                                 {
                                     Directory.Delete(path, recursive: true);
                                 }
